Simple file watcher in storage.
This will need more work, and it could be cool to add a functionality
to send only which files have been modified in the storage, instead of
sending sending the whole list of files.
For now, if a new file is added to the directory, the file list is
sent.