Check state of connections and remove those Disconnected for long enough
When counting number of files throughout the system, those in
Disconnected storage nodes are not acounted for in the total.
A storage node which has been in a disconected state long enough, will
be removed from the list in the dispatcher.
This is also usefull when the dispatcher starts after loading it's
configuration, since some storage nodes may still be online.